Fit Gaussian graphical models

Fit Gaussian graphical models using various algorithms.

fit_ggm(
  S,
  edges = NULL,
  nobs,
  K = NULL,
  iter = 10000L,
  eps = 1e-06,
  convcrit = 1,
  aux = list(),
  method = "covips",
  print = 0
)

Arguments

S

Sample covariance matrix.

edges

Generators of model; a list of integer vectors or a 2 x p matrix of integers.

nobs

Number of observations

K

Initial value of concentration matrix.

iter

Maximum number of iterations.

eps

Convergence criterion.

convcrit

Convergence criterions. See section `details`.

aux

A list of form name=value.

method

Either `"ips"` or `"fips"`.

print

Should output from fitting be printed?

Details

Convergence criterion:

* 1: max absolute difference between S and Sigmahat on edges.

* 2: difference in log likelihood divided by number of parameters in the model (number of edges + number of nodes) between successive iterations.

R-based / c++-based in combination with con / cov.

Author

Søren Højsgaard, sorenh@math.aau.dk

Examples

options("digits"=3)
data(math)

S <- cov(math)
nobs <- nrow(math)
gl <- list(1:3, 3:5)
em <- matrix(c(1,2, 2,3, 1,3, 3,4, 3,5, 4,5), nrow=2)

EPS = 1e-2

cg = fit_ggm(S, gl, nobs=nobs, eps=EPS, method="cov")
ci = fit_ggm(S, gl, nobs=nobs, eps=EPS, method="con")
rg = fit_ggm(S, gl, nobs=nobs, eps=EPS, method="cov", aux=list(engine="R"))
ri = fit_ggm(S, gl, nobs=nobs, eps=EPS, method="con", aux=list(engine="R"))

K <- solve(S)
(ci$K - K)  %>% abs %>% max
#> [1] 0.00154
(cg$K - K)  %>% abs %>% max
#> [1] 0.00154
(ri$K - K)  %>% abs %>% max
#> [1] 0.00154
(rg$K - K)  %>% abs %>% max
#> [1] 0.00154